Simplify and update build_behaviour
ohDevTools provides cleaner ways of specifying behaviour now. This makes
use of them to keep build_behaviour shorter and easier to read. This is
not intended to change any important behaviour of the build process, so
it's probably a mistake if it does. There might be slight differences in
messages output during the build, but it should still look for files in
the same places and understand the same command-line switches.